removed cops and s3 books fixtures from fixtures/webview.py and added new fixtures/cops.py file to include cops and s3 fixtures
